Restructure bridge to be more coherent.
Some folder structure is still probably needed, however this is a lot easier to find issues in now.

This is also the official commit for #3 - matrix to minecraft communications (text only). 

Since the last update, a logging system borrowed from turt2live/matrix-voyager-bot was used. This makes reading the output a lot easier.